+++ src/stimulus.rs
@@ -79,7 +79,7 @@ impl Sum for Stimulus {
 impl From<Rgb> for Stimulus {
     fn from(rgb: Rgb) -> Self {
         let prim = &rgb.space.data().primaries;
-        let rgb2xyz = rgb.observer.data().rgb2xyz(&rgb.space);
+        let rgb2xyz = rgb.observer.data().rgb2xyz(rgb.space);
         let yrgb = rgb2xyz.row(1);
         rgb.rgb
             .iter()