+++ CHANGELOG.md
@@ -18,7 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - `Gaussian` struct, to represent normal distributions as used in this library as spectral
   distributions and used for filtering.
 - `Planck` struct, with methods for calculating blackbody emission spectra.
-- Add `CieCam02` implementation.
+- Add `CieCam02` implementation, with same methods as CieCam16.
 - Add `CFI` Color Fidelity Index calculations. with `general_color_fidelity_index` and `special_color_fidelity_indices`.
 
 ### Changed
@@ -28,6 +28,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Renamed `ObserverData::xyz_fn_illuminant` to `xyz_from_fn`.
 - Renamed `ObserverData::xyz_from_std_illuminant_x_fn` to `xyz_from_colorant_fn`.
 - Replaced all instances of `&ObserverData` in the public API with `Observer`, which is an `enum`, to avoid confusion between the use of `ObserverData` and `Observer`.
+- Renamed `CieCam16::ciede2016` to `de_ucs`.
 
 ### Removed
 - Various normal distribution (gaussian) helper functions, now all collected as methods of the `Gaussian` struct.